+++ tests/e2e/cases/dynamic_require_static_buffer_concat_to_string/bcat_616263640000.ts
@@ -1,0 +1,1 @@
+module.exports.label = "bcat_long_hex";